fix: pass permission flags as Long on Android 13+

+++ app/src/main/java/com/silveseguridad/app/SecurityScanner.kt
@@ -19,7 +19,7 @@ class SecurityScanner(private val context: Context) {
 
     fun scan(): Report {
         val pm = context.packageManager
-        val packages = if (Build.VERSION.SDK_INT >= 33) pm.getInstalledPackages(PackageManager.PackageInfoFlags.of(PackageManager.GET_PERMISSIONS)) else {
+        val packages = if (Build.VERSION.SDK_INT >= 33) pm.getInstalledPackages(PackageManager.PackageInfoFlags.of(PackageManager.GET_PERMISSIONS.toLong())) else {
             @Suppress("DEPRECATION")
             pm.getInstalledPackages(PackageManager.GET_PERMISSIONS)
         }
